+++ packages/shadstack-table/src/hooks/useSST_ColumnVirtualizer.ts
@@ -35,9 +35,9 @@ export const useSST_ColumnVirtualizer = <
     () =>
       enableColumnPinning
         ? [
-            table.getLeftVisibleLeafColumns().map((c) => c.getPinnedIndex()),
+            table.getStartVisibleLeafColumns().map((c) => c.getPinnedIndex()),
             table
-              .getRightVisibleLeafColumns()
+              .getEndVisibleLeafColumns()
               .map((column) => visibleColumns.length - column.getPinnedIndex() - 1)
               .sort((a, b) => a - b),
           ]